RENDERING WORKING

+++ src/render.rs
@@ -1,4 +1,6 @@
-use bevy::prelude::*;
+use std::mem::transmute;
+
+use bevy::{ prelude::*, window::PrimaryWindow };
 
 use crate::*;
 use ray::*;
@@ -6,29 +8,26 @@ use ray::*;
 pub fn render(
   mut gizmos: Gizmos,
   query: Query<(&Transform, &FieldOfView), With<Player>>,
-  map: Res<Map>
+  window_query: Query<&Window, With<PrimaryWindow>>,
+  hits: Res<Hits>
 ) {
-  if let Ok((transform, field_of_view)) = query.single() {
+  if let Ok((transform, field_of_view)) = query.single() && let Ok(window) = window_query.single() {
     let origin = transform.translation.truncate();
-    for i in 0..field_of_view.ray_count {
-      // Get each ray's angle based on the player's rotation and the field of view
-      let angle = get_ray_angle(i, transform, field_of_view);
-      let start = transform.translation;
-      let end = start + Vec3::new(angle.cos(), angle.sin(), 0.0) * field_of_view.max_distance;
-      let ray = Ray { start: start.truncate(), sec_point: end.truncate() };
-
-      let mut nearest_hit: Option<Vec2> = None;
-      let mut nearest_dist_sq = f32::MAX;
-
-      for wall in &map.walls {
-        if let Some(hit) = ray_hit(&ray, wall) {
-          let dist_sq = origin.distance_squared(hit);
-          if dist_sq < nearest_dist_sq {
-            nearest_dist_sq = dist_sq;
-            nearest_hit = Some(hit);
-          }
+    let window_size = window.size();
+    for i in 0..RAY_COUNT {
+      let line_width = window_size.x / (RAY_COUNT as f32);
+      let iso = Isometry2d::from_xy(
+        line_width * ((i as f32) - (RAY_COUNT as f32) / 2.0) + line_width / 2.0,
+        0.0
+      );
+      let mut line_height: f32 = 0.0;
+      if let Some(hit) = hits.0[i] {
+        let dist = origin.distance(hit);
+        if dist <= field_of_view.max_distance {
+          line_height = (WALL_HEIGHT * window_size.y) / dist;
         }
       }
+      gizmos.rect_2d(iso, Vec2::new(line_width, line_height), Color::srgb(1.0, 0.0, 0.0));
       //MAKE RENDER LINES
     }
   }
